auditor/requirements.txt
fastapi  
uvicorn  
pydantic  
pikepdf  
anyio  
python-multipart  
azure-identity  
azure-core  
openai
pypdf